Dataset is generatable!
generate_dataset.py runs. Vasvi, please modify the "Convert_to_HF" folder for using Hugging Face Transformers library. NOTICE: this would not be faithful to reimplementing the baseline on Gemini models, but on a smaller model on HF that is likely to answer incorrectly so we can show the improvement works.